Add detection for scripts by shebang line
To detect if the first argument is a script file, the previous rules
are checked first: an existing file with a `.py` suffix always passes.

If that check fails but the file exists, the next stage is now to
check the shebang line and allow for any script which is a direct or
`/usr/bin/env` invocation of the following commands:
- python
- python3
- pip-run
- py
- pypy
- pypy3
- python3.*

`env -S` is allowed as well, as it may be used to pass additional
arguments to the underlying command.

A small amount of extra care is taken around handling invalid unicode
content. The first line is the only line which is decoded, and any
decoding error makes the shebang check fail.

Unit tests confirm the shebang line parsing behavior outside of the
broader context of `pip-run` invocations.