Simplify shebang check and support BOM
Rather than parsing the shebang line with shlex, only check if it
starts with `#!`. Because this simplifies reading of the line to not
really require a decode() call, take extra measures to ensure that the
shebang check supports shebang lines which start with BOM sequences
supported by python (via `codecs`).